// src/interfaces/queries.tsx
export interface QueriesListProps {
  id: string
  queriesListId: string
  title: string
}

export interface QueriesPainelProps {
  id: string
  queriesId: string
  queriesListId: string
  title: string
  subtitle: string
  image: string
  banner: string
  summary: string
  description: string
  subdescription: string
}